feat: add product page content and fix Clippy warnings
- Add "For Companies" section to README with commercial offerings
- Update GitHub Pages docs with enterprise features and CTAs
- Fix 15+ Clippy warnings in core library:
  - Use iterators instead of index-based loops
  - Apply is_multiple_of() and div_ceil() methods
  - Fix struct initialization patterns in tests
  - Implement Default trait for PSOOptions
  - Remove unnecessary casts
- Add high_strain_swarm_simulation example

+++ src/mavlink_controller.rs
@@ -283,7 +283,7 @@ impl FlightController {
             }
             MavMessage::SYS_STATUS(status) => {
                 self.state.battery_voltage = status.voltage_battery as f32 / 1000.0;
-                self.state.battery_percent = status.battery_remaining as i8;
+                self.state.battery_percent = status.battery_remaining;
             }
             _ => {}
         }
